# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Promo API",
        "description": "Promo API by PicPay",
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/consumers/{consumer_id}/cashfront": {
            "patch": {
                "tags": [
                    "cashfront"
                ],
                "summary": "Executa cashfront",
                "description": "A partir de detalhes do dep\u00f3sito e consumer, retorna a melhor campanha e cancela as similares",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/DepositCorePayload"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/consumers/{consumer_id}/transactions/{transaction_type}/{transaction_id}/undo": {
            "patch": {
                "tags": [
                    "cashback"
                ],
                "summary": "Reverte cashback e efeitos colaterais gerados por transa\u00e7\u00e3o",
                "description": "Com informa\u00e7\u00e3o de transa\u00e7\u00e3o que gerou cashback, desfaz",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "transaction_type",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/consumers/{consumer_id}/campaigns/associations": {
            "get": {
                "tags": [
                    "consumer"
                ],
                "summary": "Obt\u00e9m as as associa\u00e7\u00f5es de um usu\u00e1rio a campanhas",
                "description": "Obt\u00e9m as as associa\u00e7\u00f5es de um usu\u00e1rio a campanhas",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filtra por datas de associa\u00e7\u00e3o a partir da informada, em ISO 8601",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filtra por datas de associa\u00e7\u00e3o de fim at\u00e9 a informada, em ISO 8601",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "N\u00famero da p\u00e1gina",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Tamanho da p\u00e1gina",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Se retorna o total de itens ou n\u00e3o",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/consumers/{consumer_id}/coupon": {
            "patch": {
                "tags": [
                    "coupon"
                ],
                "summary": "Aplica cupom a usu\u00e1rio",
                "description": "Tenta aplicar cupom, que faz associa\u00e7\u00e3o a campanha",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/CouponApplicationPayload"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Consumer not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/consumers/{consumer_id}/transactions/{transaction_type}/{transaction_id}": {
            "get": {
                "tags": [
                    "cashback"
                ],
                "summary": "Obt\u00e9m transa\u00e7\u00e3o que gerou cashback",
                "description": "Com consumer e transaction ids, retorna a transa\u00e7\u00e3o",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "transaction_type",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/consumers/{consumer_id}/transactions": {
            "get": {
                "tags": [
                    "cashback"
                ],
                "summary": "Obt\u00e9m as as transa\u00e7\u00f5es que geraram cashback a um usu\u00e1rio",
                "description": "Obt\u00e9m as as transa\u00e7\u00f5es que geraram cashback a um usu\u00e1rio",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "campaign_id",
                        "in": "query",
                        "description": "Filtra por id de campanha",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "transaction_id",
                        "in": "query",
                        "description": "Filtra por id de transa\u00e7\u00e3o",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "transaction_type",
                        "in": "query",
                        "description": "Tipo p2p ou pav",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "N\u00famero da p\u00e1gina",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Tamanho da p\u00e1gina",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Se retorna o total de itens ou n\u00e3o",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/consumers/{consumer_id}/banners": {
            "get": {
                "tags": [
                    "banner"
                ],
                "summary": "Obt\u00e9m os banners de um consumer, direcionados e globais",
                "description": "Obt\u00e9m os banners direcionados a um consumer",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "consumer_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "area_code",
                        "in": "query",
                        "description": "DDD do usu\u00e1rio",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DepositCorePayload": {
            "properties": {
                "deposit": {
                    "description": "CashfrontRequest class",
                    "$ref": "#/definitions/DepositCorePayloadDetails"
                }
            }
        },
        "DepositCorePayloadDetails": {
            "properties": {
                "id": {
                    "description": "CashfrontRequest class",
                    "type": "integer",
                    "example": "34346"
                },
                "recharge_method": {
                    "description": "CashfrontRequest class",
                    "type": "string",
                    "example": "conta-corrente"
                },
                "total": {
                    "description": "CashfrontRequest class",
                    "type": "number",
                    "example": 40.62
                },
                "first_deposit": {
                    "description": "CashfrontRequest class",
                    "type": "boolean",
                    "example": true
                }
            },
            "type": "object"
        },
...
        "CouponApplicationPayload": {
            "properties": {
                "coupon": {
                    "description": "CouponRequest class",
                    "$ref": "#/definitions/CouponApplicationPayloadDetails"
                }
            }
        },
        "CouponApplicationPayloadDetails": {
            "properties": {
                "code": {
                    "description": "CouponRequest class",
                    "type": "string",
                    "example": "UBER20"
                },
                "conditions": {
                    "description": "CouponRequest class",
                    "$ref": "#/definitions/CouponApplicationPayloadConditions"
                }
            }
        },
...
    }
}